backend/requirements.txt
Flask==2.3.2
gunicorn==21.2.0
google-cloud-logging==3.8.0
google-cloud-secret-manager==2.16.1
python-dotenv==1.0.0
requests==2.31.0
SQLAlchemy==2.0.25
cloud-sql-python-connector[pg8000]==1.16.0
flask-cors==4.0.0
google-cloud-storage==3.1.0
Werkzeug==2.3.7
pytest==8.0.0
pytest-cov==4.1.0
PyJWT==2.8.0
psycopg2-binary==2.9.9
redis==5.0.1
python-jose==3.3.0
cryptography==42.0.5
Flask-Limiter==3.5.0
bcrypt==4.1.2
google-auth[requests]>=2.0.0
